Add ability to resolve measure by canonical URL
