+++ cohort-engine/src/main/java/com/ibm/cohort/engine/measure/RestFhirLibraryResolutionProvider.java
@@ -1,3 +1,8 @@
+/*
+ * (C) Copyright IBM Corp. 2020, 2020
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
 package com.ibm.cohort.engine.measure;
 
 import java.util.Map;
